+++ ANSV/naive_n2_work.h
@@ -5,6 +5,6 @@
 #ifndef IMPLEMENTATIONS_NAIVE_N_SQUARED_H
 #define IMPLEMENTATIONS_NAIVE_N_SQUARED_H
 
-std::tuple<std::array<VI, n>, std::array<VI, n>> ANSV_naive(std::array<int, n> A);
+std::tuple<std::array<VI, n>, std::array<VI, n>> ANSV_naive(std::array<long, n> A);
 
 #endif //IMPLEMENTATIONS_NAIVE_N_SQUARED_H
