+++ ANSV/seq_array_n_work.h
@@ -8,6 +8,6 @@
 #define IMPLEMENTATIONS_SEQ_ARRAY_H
 
 
-std::tuple<std::array<VI, n>, std::array<VI, n>> ANSV_seq_array(std::array<int, n> A);
+std::tuple<std::array<VI, n>, std::array<VI, n>> ANSV_seq_array(std::array<long, n> A);
 
 #endif //IMPLEMENTATIONS_SEQ_ARRAY_H
