+++ ANSV/seq_stack_n_work.h
@@ -7,6 +7,6 @@
 #ifndef IMPLEMENTATIONS_SEQ_STACK_H
 #define IMPLEMENTATIONS_SEQ_STACK_H
 
-std::tuple<std::array<VI, n>, std::array<VI, n>> ANSV_seq_stack(std::array<int, n> A);
+std::tuple<std::array<VI, n>, std::array<VI, n>> ANSV_seq_stack(std::array<long, n> A);
 
 #endif //IMPLEMENTATIONS_SEQ_STACK_H
