+++ Testing/TestANSV.h
@@ -11,8 +11,8 @@ class TestANSV {
 
 };
 
-void compareOutputOfNaiveSeqArraySeqStack(std::array<int, n> A, bool verbose=false);
+void compareOutputOfNaiveSeqArraySeqStack(std::array<long, n> A, bool verbose=false);
 
-std::tuple<std::array<VI, n>, std::array<VI, n>> testShunZhao(std::array<int, n> A);
+std::tuple<std::array<VI, n>, std::array<VI, n>> testShunZhao(std::array<long, n> A);
 
 #endif //IMPLEMENTATIONS_TESTANSV_H
